refactor: rename lp to mip.
