+++ src/main/resources/edu/kit/provideq/toolbox/mip/simple.mps
